Enhance checkout view.    - Refactor existing utility method out of views logic.    - Created a utility for views and new logic in the view.
